Extract FormStep out of ConfigForm Component
